+++ ide/app/test/ui/all.dart
@@ -1,11 +1,11 @@
-// Copyright (c) 2013, Google Inc. Please see the AUTHORS file for details.
+// Copyright (c) 2014, Google Inc. Please see the AUTHORS file for details.
 // All rights reserved. Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
 /**
  * This library defines all UI tests for Spark.
  */
-library spark.ui_tests.all_tests;
+library spark.ui.all_tests;
 
 import 'general_ui_test.dart' as general_ui_test;
